lib: change run_program() to use $PATH
GPU detection tries to launch a copy of the client using full path to
the client executable but getting the full path may fail. In this case
the code falls back to using argv[0] but argv[0] doesn't always contain
path to the client. This may happen if the client was started from shell
and the shell used $PATH to find the client executable.

Change run_program() to search $PATH for the program file if the file
name doesn't include path.