client: decode error codes in detecting GPUs
On Unixes, if GPU detection fails the client logs an error message
containing a raw value from waitpid() call. This raw value generally
requires writing an external program to decode it before it is useful
for troubleshooting.

Decode the raw value to something more useful to humans. Make a similar
change to Windows code too.