Travis: harmonize deploy scripts
On AppVeyor the archiving is done in prepare stage because they have an artefact cache for pull requests already available. So I moved the 7z stuff to prepare_deployment.sh so it does the same as the bath file. Filenames do not match because uploading to bintray is done differently but theyhave the same structure when available on bintray..