Update GAHP version string
The existing code depended on building within a git repository.  It
included the GIT_REVISION in the startup message.  Unfortunately,
if you built from a downloaded tarball, the build fails. I thought
that the version of BOINC that the GAHP was compiled against would
be useful information as it is logged.  The SVN_VERSION is still
available by using the --version command line option.  Because of
the small change in behavior, I also bumped the version number of
the GAHP to 1.0.2.