web: fix layout problem with forum tables
Forum messages are shown in a table: left column is sender info,
right column is the message.

The left column was fixed width (10em).
On very hi-res monitors people increase their font size.
This causes the "Send message" button to overflow the column.

Solution: don't use fixed width, and make the right column 100% width.
This makes the left column wide enough for its contents to fit, but no wider.

This solves the problem, but introduces a (minor) new one:
some users have extremely long user names,
and this makes the left column too wide,
not just for that post but for the whole thread.

Solution: in this particular place, if name is > 30 chars,
show only the first 30 chars followed by ellipsis.