client: don't quote the path to client in GPU detection
Client path was quoted because on Windows run_program() concatenates
arguments to a single string. Without quoting any path with a space
would cause problems interpreting the command line.

What run_program() does with its parameters is really an implementation
detail of it and users of run_program() shouldn't need to know or care
about it.

Moving quoting into run_program() also removes the need to duplicate
quoting everywhere run_program() is used.

The program name or path to it needs to be valid filesystem name and
it's always the first in argv so quoting it in run_program() is trivial.
run_program() doesn't know which other arguments may contain paths and
quoting arbitrary arguments is harder so keep quoted_data_dir for now.