db: Modified db_dump so user and host output is independent of stats_export consent.
If the project does have stats exports consent, and a user does NOT consent to stats. exports: then the total user and host in table.xml should still include the user. This is an aggregate total of the number of users and hosts that use the project.

Introduces an additional count query to count the total number of users and hosts regardless of the stats-export consent functionality.