client,mgr: pass on GUI name in <exchange_versions>
The name of GUI RPC client will be used when creating an account at a
project that requires users to consent to terms of use. The project will
register both the consent and the GUI that was used to create the
account.

Store API version as well for future use.