client: prevent a situation where a job is not throttled
I noticed an Amicable Numbers 4-CPU job wasn't being throttled,
while other jobs were.
Two types of jobs are exempt from throttling: GPU and non-CPU-intensive.
This app version was neither.
And yet it was marked as <dont_throttle/> in client_state.xml.
This is read at startup, and prevents that app version from being throttled
even if it not GPU and is not non-CPU-intensive.

I'm not sure how this situation arose.
Possibly the project had a temporary misconfiguration
where that plan class (mt) used a GPU.
This set the dont_throttle flag, and it stuck.

Anyway, the fix is to not parse <dont_throttle/> from app versions
in the client state file.
There's no need to - after parsing the app_version record,
dont_throttle is set according to whether it uses a GPU.