Travis Bintray: add attributes to versions for better cleanup
Attributes can be used to easily delete old versions. A script can check if a PR is closed and delete all versions for this PR on bintray. Or it could delete all versions older than a threshold.
Attribtes only need to be added once. Currently this is done in the Travis build because this is running for every PR.