remote job submission: add API for getting job counts
If a remote job submission system wants to maintain a constant
level of unsent jobs, it needs a way to get this value.
This commit adds a Python API for this.
It also returns other counts such as in-progress results,
and WUs needing validation/assimilation/file delete.

I could add PHP and C++ interfaces too if needed.