[Android] Add Cancel button to 'Add Project' screen
Currently there is no way to cancel 'Add Project' procedure once it is started.
This pull request add cancel button to this screen.
Also it fixes the warning with `layout_alignParentRight` setting that could work in not a proper way for right-to-left UI.
I'm not sure whether it is an issue now but definitely will be better to have.

This fixes #3136

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>