job submission: add ability to view mem/disk usage stats for a batch.
It's useful to see the stats (mean, max) of memory and disk usage,
e.g. so that you can set the appropriate bounds for future jobs.
In the case of universal VM apps (e.g. nanoHUB@home)
the natural place to put this is at the batch level;
jobs in a a batch use the same tool.

Also: add a general function for printing sizes in MB and GB,
and move a function from util.inc to the appropriate file.