[Android] Make BOINC always run foreground service
On the latest Android SDKs there are a lot of restrictions for background processes.
So if the service is run in background, it could be killed too early and BOINC will not start.
Removed setting to show notification when running only because now it makes no sense since Notification should be shown always when foreground service is running.

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>