Client: fix crashing bug in RSS feed fetch.
We were getting a pointer to an element of a vector<RSS_FEED>,
and then using it later (when HTTP op completes).
The problem is that we may have added something to the vector
in the interim, in which case the pointer may be invalid.
This is a "how did it ever work?" situation.
I guess it doesn't happen very often.

Also: remove spurious error message if AM tells us to detach from
a project we're not attached to.