Fix UB for size_t format specifiers
Using incorrect format specifier in fprintf is UB. std::size_t is the unsigned integer type of the result of the sizeof operator and can store the maximum size of a theoretically possible object of any type (including array).

C++11 imports C99 so std::printf should support the C99 %zu format specifier for size_t. The problem is that MinGW compiler doesn't support %zu.

in order to compile BOINC under MinGW we can:
1. Use preprocessor to define compiler specific macroses for printing size_t:

it can bloat public boinc headers, so not an ideal solution.

2. Define __USE_MINGW_ANSI_STDIO to 1 for using special mingw versions of printf functions that support %zu format specifier (https://sourceforge.net/p/mingw-w64/wiki2/gnu%20printf/)

Defining such macro globally may lead to unexpected changes in runtime behavior for MinGW BOINC builds, so not an option.

3. Another solution is to cast size_t to uintmax_t, which is maximum width unsigned integer type, then use format specifier for that. It involves a small performance penalty for x86 builds, since uintmax_t is 8 bytes long and cast from 4 bytes unsigned int (size_t) to 8 bytes long uintmax_t (unsigned long long) is required.

Decided to use solution #3, since it doesn't require changes in public headers and doesn't lead to review all of the printf-like code to catch runtime behavior changes of __USE_MINGW_ANSI_STDIO.