client: don't show "no work" messages as notices
On startup the client sees, for each project,
whether it's barred from using particular resources (CPU, GPU).
It was showing these situations as notices.
I think the idea was to make sure the user hadn't changed
a setting and forgotten about it.

But this was annoying overkill.
Instead, just show startup messages (not notices)
saying what (project, resource) pairs are disallowed, and wy.