configure.ac: Fix have_xss test to respect --without-x
Add test to check value of no_x to prevent undesired inclusion of libXss,
even if present. The AC_PATH_X macro sets the shell variable no_x to 'yes'
if the user gave the command line option --without-x, so this must be
respected for the screensaver idle detection. However, since the macro sets
no_x to the same value if the X Window System include files and libraries
cannot be located, provide build warning to assist with debugging.