GUI RPC client interface, linux: look for config file in /etc
When looking for gui_rpc_auth.cfg, first look in current dir,
then see if there's a file /etc/boinc-client/config.properties
with a line "data_dir=X".  If so, look in X.
Return prescriptive message in case of failure.

boinccmd: use the above, and show the message.
Note: a lot of GUI RPCs don't require the GUI RPC password.