Fixed the shutdown client dialog from showing when told not to
[Manager] Short description of change

The "shutdown connected client" modal will not show ever again if you check the "do not show this dialog again" box provided.

Detailed description of the change (if necessary)

This fix only fixes the non-persistence issue and does not allow the user to make the modal appear once again and hence reverse their check of the box. Whether to implement such facilities will be left to discussion.

Fixes #3933