client: only report a graphics app if file is there can we can execute it
Problem: when parsing an APP_VERSIOn (e.g. in scheduler reply)
we check whether it has a graphics app (i.e. a file named "graphics_app").
If so we could report it to the GUI before the file has been downloaded,
causing an error when "Shot Graphics" is clicked.

Also, on Mac we check whether the graphics app can run on
this CPU, given its header.
But this check was being done during APP_VERSION parsing,
when typically the file doesn't exist yet.

Fix: when parsing APP_VERSION,
store a pointer to the FILE_INFO for the graphics app.
When reporting an ACTIVE_TASK to the GUI,
check whether the file is present and (on Mac) can be executed.
Only then tell the GUI about it.
If the file download failed or the file can't be executed,
zero out the pointer so we don't do the check again.