[Android] Fix Tasks loading
On modern devices with more than 4 cores or when there is a big tasks cache set could happen a situation when 0 tasks are shown.
This happens because of the limit of 1MB to transfer data between processes.
This commit breaks this flow and instead adds a possibility to load data by chunks.
Also small optimization implemented to update only active tasks every second.
Full update is done every 10 seconds now because most of the time non-active tasks are not updated so often.

This fixes #2123

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>