Suggested code
No need to protect from android 7 or above, because it just add more suggested folders to look at it.
Test this code on my armv6 - Galaxy gio device with android 4.4 (cyanogen os). It not recognize my gpu. (there isn't opencl for Adreno 200) and it running and crunching fine without crashing.