client: be conservative about non-BOINC CPU usage on underflow
Instead of assume that the non-BOINC CPU usage was 0%, assume that it
was 50%. This potentially prevents BOINC from kicking-in in the
presence of short-lived processes (if configured to run only if
non-BOINC CPU usage is below 50%).

The main issue is that BOINC's calculation of non-BOINC CPU usage time
is incorrect, as processes which exited within the sample
interval (typically 10 seconds) are not taken into consideration. This
leads to BOINC underestimating the non-BOINC CPU usage, especially in
situations where a lot of short-lived processes are present. This
commit does not fix that, but mitigates the effects of this. For more
information, see issue #4327.