client: fix work-fetch logic when max concurrent limits are used
The round-robin simulation would stop simulating jobs for a project
once a max concurrent limit (app or project) was reached.
As a result it would decide there was a shortfall,
and keep requesting work up to the limit of 1000 jobs.

To fix this:
1) keep simulating a project after an MCL is reached
2) for each (project, resource) pair, keep track of the latest
    simulation time T when an MCL was reached.
3) for such a project, don't request work for a resource if
    T > now + work buf size

This allows us, e.g., to request GPU jobs from a project
even if its CPU jobs (taking MCL into account) fill the buffer.

This works in the simulation case that showed the problem (#192).

Also: add a bit more logging, and improve names