configure.ac: Do not probe for <avxintrin.h>
With GCC 12, this header file cannot be included on its own.
Doing so fails with #error, but also produces many warnings
related to implicit ints and implicit function declarations.
This is inconvenient for distributions who are trying to
switch to newer C versions (without implicit ints/function
declarations), and monitor builds for suspicious warnings
related to obsolete C features.