client: accept prefs from project if using local prefs
Recall: the client has
- web prefs: from a project or AM.  global_prefs.xml
- local prefs: global_prefs_override.xml

Prefs from an AM have priority over prefs from projects.
If the client is using an AM and has prefs from it,
prefs from a project (in a scheduler reply) are silently ignored.

Change things so that prefs from a project are accepted
(written to global_prefs.xml) if you're using an AM but are
using local prefs.
So if you're using an AM and want prefs from a project P:
- switch to local prefs
- update P
- switch to web prefs (which will now be from P).

This is admittedly awkward.