web: fix problems with download page on project sites
The Linux installers on the BOINC web site (.sh files)
have out-of-date binaries, and are not real installers.
We shouldn't direct users to them.
But projects and AMs that use auto-attach were directing
Linux users to these installers.

This change fixes that.
Linux users are not shown a download button, but are instead directed to
https://boinc.berkeley.edu/wiki/Installing_on_Linux
which lists various options for installing BOINC on Linux
(starting with package managers).

Also: the download page (download_software.php) is also used
by account managers (Science United).
In this case, it should show instructions for
adding an account manager, not attaching to a project.

Note: there's an analogous update to download.php on the BOINC web site:
if the machine is Linux,
show a link to the Linux instructions rather than a download button.