Win: comment out (currently) unneeded fix
The VS header files have a conflict:
various structures (such as STRING) are defined differently
in ntsecapi.h than in winternl.h

This error appears off and on, for reasons I don't understand.
Changes in lib/str_util.h seem to trigger it.
Earlier in this PR is was happening.
I found a workaround on stackoverflow.com and added it.

But now the error is not happening, even without the workaround.
So, for now, I'm commenting out the workaround.
It's there if we need it in the future.