web-based remote job submission: improve interface for getting output files
Problem: the old interface for getting output files (get_output.php)
downloaded files with their physical name,
which is a cryptic string with no extension.
If the output file is, for example, a .zip, you don't see this.
You also don't see the logical name of the file.

Solution: give downloaded files better names.
For example, if a result has a single output file
with logical name 'output.zip',
download it as resultname__output.zip
Similar for multiple output files per result,
for workunits, for single files from a result, and for batches.

Fixes #5262

This can be used as a web API, or directly from web pages (submit.php)
This replaces the old interface,
but I'm leaving it there in case anyone is using it.

Add utility functions for
- getting the full path of an output file
- getting the logical names of a result's output files