[Android] Revert back to target API 21
In #3418 I changed the behavior of BOINC to run always in the foreground mode.
This lead to the situation when notification was always shown, even when BOINC in not running, and thus making a lot of notification noise.
3 years ago I thought that this is good decision and a necessary evil.
Unfortunately, I was wrong.
My main motivation of this was to be able to be on track, and have our application in the Play Store.
But every new API increase the limitations from the Android side become stronger and stronger, and prevent us from doing what we want.
For example, in the Android versions starting from Oreo, it is not possible to start a service from the background.
And while we can leave with that fact, we received and issue with the notifications.
I thought, that if we will have BOINC always running on the foreground, it will be possible to keep it always running and prevent it from being killed by the system.
Unfortunately, this is not the case.
The system still kills the application, and the notification is still shown.
And this is not the only problem.
Other problem is that starting from API 29, we are not able to start executables downloaded from the server.
Since this is a core functionality, we are not able to target API 29 or higher.
And since we are not able to target API 29 or higher, we are not able to put BOINC in the Play Store.
And thus it makes no sense to target any API higher than 21 because it will just kill the functionality of the application.
So, I decided to revert back to the API 21, and make the application work as it was before.
But even in this case there is a chance that on the newer version of Android, some of the gard limitations will be applied, and it will prevent (again) BOINC from working.
But at least currently application is working, and we can add necessary features to it.
The only issue I see now is that the user will see a notification, that this application was developed for the older version of Android and thus might not work correctly,
but I believe we can live with that.

This fixes: #4189, #4218 and #4190

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>