server (all components that use DB): don't use MYSQL_OPT_RECONNECT if deprecated
Starting with 8.0.34, the MySQL reconnection option is deprecated,
and using it produces a warning message.
So don't use it with MySQL 8.0.34 and later.

Note: this means that if the MySQL server dies,
long-running programs like the transitioner will stop working,
and you'll have to stop/start the project.
If this proves to be a problem, we could add our own reconnect logic.