client: handle changes to project master URLs
The Einstein@home web site says its master URL is
https://einsteinathome.org.
But its scheduler replies say the master URL is
https://einstein.phys.uwm.edu
If you attach to URL1 the client creates a file
account_einsteinathome.org.xml
Then it does a scheduler RPC and gets a reply saying the master URL is URL2.
It then creates a file
account_einstein.phys.uwm.edu.xml

If you then detach from the project, it deletes the 2nd account file
but the first one is still there,
so the next time you run the client you're still attached to Einstein@home!

To fix this situation:

1) on startup, check for misnamed account files
(file name doesn't match master URL in the file).
Delete (if there's already a file of the right name) or rename.

2) on scheduler RPC reply with a URL that differs by more than http(s)://,
rename the account file and the project directory.
Reset the project (remove jobs and other info)
since soft links point to the old project dir.