client: fix non-CPU-intensive logic
Problem: If a project is non-CPU-intensive and has no jobs
we ask it for work repeatedly, without backoff.
A project is marked as non-CPU-intensive if
it has no CPU-intensive apps.
This means that a project with no apps
(e.g. because it doesn't have any jobs)
is marked as non-CPU-intensive and pinged without backoff.

Solution: mark a project as non-CPU-intensive only if its
scheduler reply says it is
(i.e. there's <non_cpu_intensive/> in its config.xml).