Wrapper (Unix): use SIGSTOP instead of SIGTSTP to suspend subprocesses
The former can't be caught, the latter can.
The goal was to support wrapped program that are themselves wrappers of some sort.

The problem (as reported in #5840) is that, contrary to documentation,
SIGTSTP doesn't actually stop processes not connected to a TTY
(at least on Linux - I checked).

This PR changes the default signal back to SIGSTOP,
but provides the option (via a --use_tstp cmdline arg to wrapper)
to use SIGTSTP instead.