client: make the --app_test feature more general.
This feature lets you run an app under a BOINC client
without having a project.
You describe the app (i.e. its input/output files) in C++ code,
in the function app_test_init().
I used this to debug sporadic apps and VM aps.

This PR is a rewrite of app_test_init() to make it clearer,
and to facilitate describing apps with input and output files.