client: if GPU is missing, discard app versions and results that refer to it.
This addresses an issue introduced when we changed GPU names from
(e.g.) 'Apple M3 Pro' to 'apple_gpu'.

Note: this means that if a host has in-progress jobs using a discrete GPU,
and you remove that GPU, those jobs will be discarded.
This is a change, which I think is good because otherwise
you'd get error messages forever.