web: make buttons more legible
The bootstrap 'btn-success' colors are white text on light green.
This is hard to read.
Change this to a darker green (seagreen).
There are two cases: form submit buttons, and links.
Lots of each; not sure I got them all.
Each of these should be factored into a single function; maybe later.

Also:
- use sprintf() instead of "foo".tra("xxx")."blah"
- fix sending HTML email with linux 'mail' command
- fix display of friends