clarify meaning of DEBUG in C++ code
I was testing boinc_cmd and was surprised to see that it was
printing the GUI RPC XML messages.
Turned out this was because of a puts(req) conditioned on DEBUG,
which is set (globally) when you configure with --enable-debug.
I looked around and there were several other cases of
very specific debugging printf's enabled by DEBUG.

This isn't the right idea.
A global flag should do global things (like compile with -g).
Use specific #defines (like SHOW_MSGS for GUI RPCs).