[Windows] Detect WSL version of enabled distros.
In the most cases, WSL2 (currently the latest) is the recommended version, but there are some cases when WSL1 is more suitable.
For example, WSL1 is faster in some cases (e.g. file operations), and it is the only option for Windows 10 Home users.
More details here: https://docs.microsoft.com/en-us/windows/wsl/compare-versions

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>